refactor ShopView
