+++ PlantApp/Views/Shop View/ShopViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  ShopViewModel.swift
+//  PlantApp
+//
+//  Created by Aeliana Shen on 3/21/24.
+//
+
+import Foundation
+import SwiftUI
+
+class ShopViewModel: ObservableObject {
+    @Published var searchText: String = ""
+    
+    func searchPlants() { }
+}
