--- PlantApp/Views/ShopView.swift
@@ -1,36 +1,0 @@
-//
-//  ShopView.swift
-//  PlantCodeiOS15
-//
-//  Created by Aeliana Shen on 6/6/23.
-//
-
-import SwiftUI
-
-struct ShopView: View {
-    var body: some View {
-        VStack(spacing: 0.0) {
-            Text("Plants")
-                .fontWeight(.bold)
-                .font(.title2)
-                .foregroundColor(ProjColor.PrimaryGreen)
-            
-            SearchBar(searchText: .constant(""))
-                .padding(EdgeInsets(top: 25, leading: 20, bottom: 20, trailing: 20))
-            
-            ScrollView(.vertical, showsIndicators: false) {
-                VStack {
-                    ScrollSection()
-                }
-            }
-            
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .top)
-    }
-}
-
-struct ShopView_Previews: PreviewProvider {
-    static var previews: some View {
-        ShopView()
-    }
-}
